requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-filter==25.1
djangorestframework_simplejwt==5.5.0
celery==5.5.2
redis==6.2.0
channels~=4.2.2
channels_redis==4.2.1